95% norminetted
