+++ src/textures.c
@@ -1,3 +1,14 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   textures.c                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: smischni <smischni@student.42wolfsburg.de> +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/12/14 10:45:02 by smischni          #+#    #+#             */
+/*   Updated: 2022/12/14 10:45:38 by smischni         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
 #include "../cub3d.h"
 
@@ -86,15 +97,15 @@ int	get_texture_color_sprite(t_img *texture, int x, int y)
 t_img	*identify_texture(t_data *data)
 {
 	t_img	*texture;
-	
-	texture = NULL;//can it happen that none of these apply?
-	if (data->cur_ray->dir.y < 0 && data->cur_ray->ori == 1) // N
+
+	texture = NULL;
+	if (data->cur_ray->dir.y < 0 && data->cur_ray->ori == 1)
 		texture = &(data->N_texture);
-	else if (data->cur_ray->dir.y > 0 && data->cur_ray->ori == 1) // S
+	else if (data->cur_ray->dir.y > 0 && data->cur_ray->ori == 1)
 		texture = &(data->S_texture);
-	else if (data->cur_ray->dir.x > 0 && data->cur_ray->ori == 0) // E
+	else if (data->cur_ray->dir.x > 0 && data->cur_ray->ori == 0)
 		texture = &(data->E_texture);
-	else if (data->cur_ray->dir.x < 0 && data->cur_ray->ori == 0) // W
+	else if (data->cur_ray->dir.x < 0 && data->cur_ray->ori == 0)
 		texture = &(data->W_texture);
 	return (texture);
 }
