+++ src/utils.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   utils.c                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: smischni <smischni@student.42wolfsburg.de> +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/12/14 10:45:57 by smischni          #+#    #+#             */
+/*   Updated: 2022/12/14 10:46:15 by smischni         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "../cub3d.h"
 
 /**
@@ -50,5 +62,5 @@ void	ft_error(char *msg, t_data *data)
 		data = NULL;
 	// if (data)
 	// 	free_all_shit(data);
-	exit(0);//OPEN: any reason for another exit code?
+	exit(0);
 }
